Move some bits of the shake file into separate modules
