+++ support/shake/1lab-shake.cabal
@@ -37,7 +37,14 @@ executable shake
       unordered-containers,
       uri-encode
     hs-source-dirs:   app
-    other-modules: Agda HTML.Backend, HTML.Base
+    other-modules:
+        Agda
+      , HTML.Backend
+      , HTML.Base
+      , Shake.Git
+      , Shake.KaTeX
+      , Shake.LinkGraph
+      , Shake.Markdown
     default-language: Haskell2010
     ghc-options: -Wextra -Wall -Wno-name-shadowing -Wno-implicit-prelude
