+++ support/shake/app/HTML/Backend.hs
@@ -1,7 +1,7 @@
 -- Copyright (c) 2005-2021 remains with the Agda authors. See /support/shake/LICENSE.agda
 
 -- | Backend for generating highlighted, hyperlinked HTML from Agda sources.
-{-# LANGUAGE DeriveGeneric, DeriveAnyClass, FlexibleContexts, ViewPatterns #-}
+{-# LANGUAGE FlexibleContexts, ViewPatterns #-}
 module HTML.Backend
   ( htmlBackend
   , builtinModules
